# PaperTrail-Angular-GUI/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "PaperTrail BookShare OpenAPI specification",
        "description": "OpenApi documentation for PaperTrail BookShare API",
        "termsOfService": "https://www.papertrailbookshare.com/terms",
        "contact": {
            "name": "Dhiraj",
            "url": "https://www.linkedin.com/in/dhiraj-birajdar-pune/",
            "email": "papaertrail.bookshare@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        },
        {
            "url": "https://www.papertrailbookshare.com",
            "description": "Production server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/feedbacks": {
            "post": {
                "tags": [
                    "Feedback Controller"
                ],
                "operationId": "saveFeedback",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "findAllBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "saveBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/cover/{book-id}": {
            "post": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "uploadBookCover",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "description": "Cover image file",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrow/{book-id}": {
            "post": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "borrowBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/shareable/{book-id}": {
            "patch": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "updateShareable",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/return/{book-id}": {
            "patch": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "returnBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/return/approve/{book-id}": {
            "patch": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "approveReturnedBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/archived/{book-id}": {
            "patch": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "updateArchivedStatus",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/feedbacks/book/{book-id}": {
            "get": {
                "tags": [
                    "Feedback Controller"
                ],
                "operationId": "findAllFeedbacksByBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{book-id}": {
            "get": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "getBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/returned": {
            "get": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "findAllReturnedBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/owner": {
            "get": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "findAllBooksByOwner",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrowed": {
            "get": {
                "tags": [
                    "Book Controller"
                ],
                "operationId": "findAllBorrowedBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FeedbackRequest": {
                "required": [
                    "bookId",
                    "comment"
                ],
                "type": "object",
                "properties": {
                    "note": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "number",
                        "format": "double"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "bookId": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "BookRequest": {
                "required": [
                    "author",
                    "isbn",
                    "synopsis",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "synopsis": {
                        "type": "string"
                    },
                    "shareable": {
                        "type": "boolean"
                    }
                }
            },
...
            "FeedbackResponse": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "number",
                        "format": "double"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "ownFeedback": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseFeedbackResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FeedbackResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "BookResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "synopsis": {
                        "type": "string"
                    },
                    "owner": {
                        "type": "string"
                    },
                    "cover": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "byte"
                        }
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "shareable": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseBookResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "BorrowedBookResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "returned": {
                        "type": "boolean"
                    },
                    "returnApproved": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseBorrowedBookResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BorrowedBookResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}