# PaperTrail-SpringBoot-RestAPI/src/main/java/com/papertrail/book/BorrowedBookResponse.java
package com.papertrail.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Integer id;
    private String title;
    private String author;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}