# PaperTrail-SpringBoot-RestAPI/src/main/java/com/papertrail/feedback/FeedbackResponse.java
package com.papertrail.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}